Add option to exclude fields or dialogs
This disregards the element and never even writes it into the config.
Intended for when there's some field that the user changed by accident.

The usefulness of excluding buttons (which excludes their dialogs) is
questionable, but I guess if a dialog opened up when recording that
wouldn't open up otherwise (??) this could be used.