GitSigns: Implement "One Fat Diff" option
Depending on the original file it was possible for the diff to fail.
This happened specifically when a chunk's anchor line was over 80
characters long. In this case git's unified diff output shortens the
chunk's anchor line.

Since the script was searching for a subtitle line identical to the
chunk's anchor line, this would fail.

This introduces an option to bypass looking for the anchor line by
having the diff just be one huge chunk instead of invididual chunks per
change (±2 lines).

Note we're intentionally not saving the one_fat_diff variable because
it's supposed to be an intentional choice, not a default when it was
used once.